Merged master changes into branch and moved NON-MONITORED version code into this branch
The master went through a number of changes during the lifespan of this
branch so far. I merged these changes into this branch, with hopes that
this branch would now build correctly. I have also merged the code from
my un-monitored version of this branch in accordance with this. I have
been using an unmonitored version of the master--an older
version--because after restructuring, the master stopped building
correctly, but it has now changed and I think that this issue has been
resolved (hopefully)