Credit to Kert for save.cpp line
I would like to give particular thanks to Kert for discovering that
save.cpp was not actually saving the scrollSpell data to the XML tree,
formatted by the XSD file. He even provided the necessary line and
syntax for navigating the tree to save the data in the proper location.
I also added some comments explaining the structure of the XSD file in
this case and how it may differ from the hero class structure.