Adding back Ironfist Project Filters
Somehow, this branch lost almost all of the Filters for the Ironfist
Project. I took the VCXPROJ and \FILTERS file from the master and then
added in the new files that came with this branch to the filters
manually. All is as it should be, with the exception of the name of the
Filters in both projects that represent the "spell" folder; they are
named "spells", and should probably be named "spell", which I will
correct in the next commit