Performed initial task investigation and fixed editManager class name
Explored HEROES2W and EDITOR2W code and played with editor random map
generator.

Marked parts of code in comments for later investigation.

Determined that the editor would not be very relevant to this task, due
to lack of placement of creature/faction-specific objects, other than
the elemental shrines.

Corrected definition of editManager::GenerateRandomMap, as the class
name was previously missing an "r".