setNumberOfCreature
The lua script can't recognize the data type town-object as a variable however GetCurrentTown() can be passed as the first parameter to this function setNumberOfCreatures. GetCurrentTown() only works if it is within OnTownOpen(town) callback so in order to use OnNewDay() to invoke setNumberOfCreatures, setNumberOfCreatures should be within in OnTownOpen and a global Boolean used in both OnNewDay and OnTownOpen. The state of the global Boolean can determine when SetNumberOfCreatures() is called.