Notes, pre-emptive allocations refined, profiling
Notes on certain issues and problems

Refined some allocations; removed game-breaking allocations, fixed some
poorly calculated sizes.

Profiling to identify what is taking the majority of the time during
saving. Unfortunately, it seems that the ironfist_save::save() function
is taking up 95% of save time, and this is defined in an auto-generated
file. Further discussion will be needed to proceed further.